+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+### Changes (2023-04-04)
+
+Added more PB-WD mappings.
+
+#### Added
+
+- Event nodes with PB mappings on the first tab [here](https://docs.google.com/spreadsheets/d/1ThN8TEzUeGb8_Oun5E-Bpin1aSAAWdgvAzlUyXcedzs/edit?usp=sharing)
+
+#### Changed
+
+- Mappings for DWD nodes on the second tab [here](https://docs.google.com/spreadsheets/d/1ThN8TEzUeGb8_Oun5E-Bpin1aSAAWdgvAzlUyXcedzs/edit?usp=sharing)
+- Corrected PB mapping in DWD_Q26972
+
 ### Changes (2023-03-31)
 
 Fixing various small issues.